Enum rowdy::auth::Error

pub enum Error {
    AuthenticationFailure,
    GenericError(String),
    HyperError(Error),
    MissingAuthorization {
        realm: String,
    },
}

Authentication errors

Variants
AuthenticationFailure

Authentication was attempted successfully, but failed because of bad user credentials, or other reasons.

GenericError(String)

A generic error

HyperError(Error)

An error due to hyper, such as header parsing failure

MissingAuthorization

The Authorization HTTP request header was required but was missing. This variant will respond with the appropriate WWW-Authenticate header.

Fields of MissingAuthorization

realm: String

The HTTP basic authentication realm