Struct rowdy::auth::simple::SimpleAuthenticator

pub struct SimpleAuthenticator { /* fields omitted */ }

A simple authenticator that uses a CSV backed user database. DO NOT USE THIS IN PRODUCTION

Requires the simple_authenticator feature, which is enabled by default.

The user database should be a CSV file, or a "CSV-like" file (meaning you can choose to use some other character as field delimiter instead of comma) where the first column is the username, the second column is a hashed password, and the third column is the salt.

Password Hashing

See SimpleAuthenticator::hash_password for the implementation of password hashing. The password is hashed using the argon2i algorithm with a randomly generated salt.

Methods

impl SimpleAuthenticator

pub fn new<R: Read>(csv: Reader<R>) -> Result<Self, Error>

Create a new SimpleAuthenticator with the provided a CSV Reader.

pub fn with_csv_file(
path: &str,
has_headers: bool,
delimiter: u8
) -> Result<Self, Error>

Create a new SimpleAuthenticator with a path to a CSV file.

pub fn hash_password(password: &str, salt: &[u8]) -> Result<String, Error>

Hash a password with the salt. See struct level documentation for the algorithm used.

pub fn verify(
&self,
username: &str,
password: &str,
include_refresh_payload: bool
) -> Result<AuthenticationResult, Error>

Verify that some user with the provided password exists in the CSV database, and the password is correct.

Returns the payload to be included in a refresh token if successful