Struct rowdy::auth::simple::SimpleAuthenticatorConfiguration

pub struct SimpleAuthenticatorConfiguration {
    pub csv_path: String,
    pub has_headers: bool,
    pub delimiter: char,
}

(De)Serializable configuration for SimpleAuthenticator. This struct should be included in the base Configuration.

Examples

{
    "csv_path": "test/fixtures/users.csv",
    "has_headers": false,
    "delimiter": " "
}

Fields
csv_path: String

Path to the CSV database, in the format described by SimpleAuthenticator. This should be relative to the working directory, or an absolute path

has_headers: bool

Whether the CSV file has a header row or not. Defaults to false.

delimiter: char

The delimiter char. By default, this is ','. Because of the limitation of the CSV crate which elects to only allow delimiters with one byte, the char will be truncated to only one byte. This means you should only use delimiters that are ASCII.