Struct rowdy::auth::Authorization

pub struct Authorization<S: Scheme + 'static>(pub Authorization<S>);

Authorization HTTP Request Header

Methods

impl<S: Scheme + 'static> Authorization<S>

fn new<'a>(header: &'a str) -> Result<Self, Error>

Create a new Authorization header

fn is_basic(&self) -> bool

Convenience function to check if the Authorization is Basic

fn is_bearer(&self) -> bool

Convenience function to check if the Authorization is Bearer

fn is_string(&self) -> bool

Convenience function to check if the Authorization is None

impl Authorization<Basic>

fn username(&self) -> String

Convenience method to retrieve the username from a HTTP Basic Authorization request header

fn password(&self) -> Option<String>

Convenience method to retrieve the password from a HTTP Basic Authorization request header

impl Authorization<Bearer>

fn token(&self) -> String

Convenience method to retrieve the token from a bearer Authorization request header.

impl Authorization<String>

fn string(&self) -> String

Convenience method to retrieve the token from an arbitrary Authorization request header.