Trait rowdy::auth::Authenticator

pub trait Authenticator<S: Scheme + 'static>: Send + Sync {
    fn authenticate(
        &self, 
        authorization: &Authorization<S>, 
        include_refresh_payload: bool
    ) -> Result<AuthenticationResult, Error>;

    fn authenticate_refresh_token(
        &self, 
        _payload: &JsonValue
    ) -> Result<AuthenticationResult, Error> { ... }

    fn prepare_authentication_response(
        &self, 
        authorization: &Authorization<S>, 
        request_refresh_token: bool
    ) -> Result<AuthenticationResult, Error> { ... }

    fn prepare_refresh_response(
        &self, 
        refresh_payload: &JsonValue
    ) -> Result<AuthenticationResult, Error> { ... }
}

Authenticator trait to be implemented by identity provider (idp) adapters to provide authentication. Each idp may support all the schemes supported, or just one.

Usually, you will want to include an Authenticator trait object as part of Rocket's managed state. Before you can do that, however, you will need to Box it up. See example below.

Examples

No-op authenticator

You can refer to the source code for the NoOp authenticator for a simple implementation.

Simple Authenticator

Refer to the MockAuthenticator[../../src/rowdy/auth/mod.rs.html] implemented in the test code for this module.

Required Methods

fn authenticate(
&self,
authorization: &Authorization<S>,
include_refresh_payload: bool
) -> Result<AuthenticationResult, Error>

Verify the credentials provided in the headers with the authenticator for the initial issuing of Access Tokens.

If the Authenticator supports re-issuing of access tokens subsequently using refresh tokens, and it is requested for, the function should return a JsonValue containing the payload to include with the refresh token.

Users should not userauthenticate directly and use prepare_authentication_response instead.

Provided Methods

fn authenticate_refresh_token(
&self,
_payload: &JsonValue
) -> Result<AuthenticationResult, Error>

Verify the credentials provided with the refresh token payload, if supported by the authenticator.

A default implementation that returns an Err(::Error::UnsupportedOperation) is provided.

Users should not user authenticate directly and use prepare_refresh_response instead.

fn prepare_authentication_response(
&self,
authorization: &Authorization<S>,
request_refresh_token: bool
) -> Result<AuthenticationResult, Error>

Prepare a response to an authentication request by first verifying credentials. If validation fails, will return an Err with the response to be sent. Otherwise, the unwrapped authentication result will be returned in an Ok. This function will also check that the authenticator behaves correctly by checking that it does not return a refresh token payload when it is not requested for

fn prepare_refresh_response(
&self,
refresh_payload: &JsonValue
) -> Result<AuthenticationResult, Error>

Prepare a response to a refresh request by first verifying the refresh payload.

If validation fails, will return an Err with the response to be sent. Otherwise, the unwrapped authentication result will be returned in an Ok. This function will also check that the authenticator behaves correctly by checking that it does not return a refresh token payload