Trait rowdy::auth::AuthenticatorConfiguration

pub trait AuthenticatorConfiguration<S: Scheme + 'static>: Send + Sync + Serialize + DeserializeOwned + 'static {
    type Authenticator: Authenticator<S>;
    fn make_authenticator(&self) -> Result<Self::Authenticator, Error>;
}

Configuration for the associated type Authenticator. [rowdy::Configuration] expects its authenticator field to implement this trait.

Before launching, rowdy will attempt to make an Authenticator based off the configuration by calling the make_authenticator method.

Associated Types

type Authenticator: Authenticator<S>

The Authenticator type this configuration is associated with

Required Methods

fn make_authenticator(&self) -> Result<Self::Authenticator, Error>

Using the configuration struct, create a new Authenticator.