Module rowdy::auth::util

Utility functions to aid in managing users

Primarily, this module provides a useful function to hash a password, in combination with a salt to produce a hash with Argon2i.

The hash produced will be 32 bytes long.

Functions

generate_salt
Generate a new random salt based on the configured salt length

hash_password_digest
Given a password and a salt, generate an argon2i hash 32 bytes in length

hex_dump
Dump a bunch of bytes as a hexadecimal string