Function rowdy::auth::util::generate_salt

pub fn generate_salt(salt_length: usize) -> Result<Vec<u8>, ()>

Generate a new random salt based on the configured salt length

For argon2i, you should use a salt between 8 and 232-1 bytes

If this function fails, no extra details can be provided. See Unspecified