Enum rowdy::token::Error

pub enum Error {
    TokenAlreadyEncoded,
    TokenAlreadyDecoded,
    TokenNotEncoded,
    TokenNotDecoded,
    NoRefreshToken,
    RefreshTokenAlreadyEncrypted,
    RefreshTokenAlreadyDecrypted,
    RefreshTokenNotDecrypted,
    RefreshTokenNotEncrypted,
    InvalidService,
    InvalidIssuer,
    InvalidAudience,
    GenericError(String),
    IOError(Error),
    JWTError(Error),
    TokenSerializationError(Error),
}

Token errors

Variants
TokenAlreadyEncoded

Raised when attempting to encode an already encoded token

TokenAlreadyDecoded

Raised when attempting to decode an already decoded token

TokenNotEncoded

Raised when attempting to use a decoded token when an encoded one is expected

TokenNotDecoded

Raised when attempting to use an encoded token when an decoded one is expected

NoRefreshToken

Raised when attempting to perform an operation on the refresh token, but the refresh token is not present

RefreshTokenAlreadyEncrypted

Raised when attempting to encrypt and sign an already encrypted and signed refresh token

RefreshTokenAlreadyDecrypted

Raised when attempting to decrypt and verify an already decrypted and verified refresh token

RefreshTokenNotDecrypted

Raised when attempting to use an encrypted refresh token when a decrypted one is expected

RefreshTokenNotEncrypted

Raised when attempting to use an decrypted refresh token when a encrypted one is expected

InvalidService

Raised when the service requested is not in the list of intended audiences

InvalidIssuer

Raised when the issuer is invalid

InvalidAudience

Raised when the audience is invalid

GenericError(String)

Generic Error

IOError(Error)

IO Error when reading keys from files

JWTError(Error)

Errors during token encoding/decoding

TokenSerializationError(Error)

Errors during token serialization