Enum rowdy::token::Secret

pub enum Secret {
    None,
    ByteSequence(ByteSequence),
    Bytes {
        path: String,
    },
    RSAKeyPair {
        rsa_private: String,
        rsa_public: String,
    },
}

Secrets for use in signing and encrypting a JWT. This enum (de)serialized as an untagged enum variant.

Defaults to None.

Serialization Examples

No secret

{
    "secret": null
}

extern crate rowdy;
#[macro_use]
extern crate serde_derive;
extern crate serde_json;

use rowdy::token;

#[derive(Serialize, Deserialize)]
struct Test {
    secret: token::Secret
}

let json = r#"{ "secret": null }"#;
let deserialized: Test = serde_json::from_str(json).unwrap();

HMAC secret string

{
    "secret": "some_secret_string"
}

extern crate rowdy;
#[macro_use]
extern crate serde_derive;
extern crate serde_json;

use rowdy::token;

#[derive(Serialize, Deserialize)]
struct Test {
    secret: token::Secret
}

let json = r#"{ "secret": "some_secret_string" }"#;
let deserialized: Test = serde_json::from_str(json).unwrap();

RSA Key pair

{
    "secret": { "rsa_private": "private.der", "rsa_public": "public.der" }
}

extern crate rowdy;
#[macro_use]
extern crate serde_derive;
extern crate serde_json;

use rowdy::token;

#[derive(Serialize, Deserialize)]
struct Test {
    secret: token::Secret
}

let json = r#"{ "secret": { "rsa_private": "private.der", "rsa_public": "public.der" } }"#;
let deserialized: Test = serde_json::from_str(json).unwrap();

Variants
None

No secret -- used when no signature or encryption is required.

ByteSequence(ByteSequence)

Secret for HMAC signing

Bytes

Path to a file containing the byte sequence for HMAC signing or encryption key

Fields of Bytes

path: String

Path to the file containing the byte sequence for a HMAC signing or encryption key

RSAKeyPair

DER RSA Key pair. See [jwt::jws::Secret] for more details.

Fields of RSAKeyPair

rsa_private: String

Path to DER encoded private key

rsa_public: String

Path to DER encoded public key