Struct rowdy::token::Configuration

pub struct Configuration {
    pub issuer: StringOrUri,
    pub allowed_origins: AllOrSome<HashSet<Url>>,
    pub audience: SingleOrMultiple<StringOrUri>,
    pub signature_algorithm: Option<SignatureAlgorithm>,
    pub secret: Secret,
    pub expiry_duration: Duration,
    pub refresh_token: Option<RefreshTokenConfiguration>,
}

Token configuration. Usually deserialized as part of [rowdy::Configuration] from JSON for use.

Examples

This is a standard JSON serialized example.

{
    "issuer": "https://www.acme.com",
    "allowed_origins": { "Some": ["https://www.example.com", "https://www.foobar.com"] },
    "audience": ["https://www.example.com", "https://www.foobar.com"],
    "signature_algorithm": "RS256",
    "secret": {
                "rsa_private": "test/fixtures/rsa_private_key.der",
                "rsa_public": "test/fixtures/rsa_public_key.der"
               },
    "expiry_duration": 86400
}

extern crate rowdy;
extern crate serde_json;

use rowdy::token::Configuration;

let json = r#"{
    "issuer": "https://www.acme.com",
    "allowed_origins": { "Some": ["https://www.example.com", "https://www.foobar.com"] },
    "audience": ["https://www.example.com", "https://www.foobar.com"],
    "signature_algorithm": "RS256",
    "secret": {
                "rsa_private": "test/fixtures/rsa_private_key.der",
                "rsa_public": "test/fixtures/rsa_public_key.der"
               },
    "expiry_duration": 86400
}"#;
let deserialized: Configuration = serde_json::from_str(json).unwrap();

Variations for the fields allowed_origins, audience and secret exist. Refer to their type documentation for examples.

Fields
issuer: StringOrUri

The issuer of the token. Usually the URI of the authentication server. The issuer URI will also be used in the UUID generation of the tokens, and is also the realm for authentication purposes.

allowed_origins: AllOrSome<HashSet<Url>>

Origins that are allowed to issue CORS request. This is needed for browser access to the authentication server, but tools like curl do not obey nor enforce the CORS convention.

audience: SingleOrMultiple<StringOrUri>

The audience intended for your tokens. The service request paremeter will be validated against this

signature_algorithm: Option<SignatureAlgorithm>

Defaults to none

secret: Secret

Secrets for use in signing a JWT. This enum (de)serialized as an untagged enum variant.

Defaults to None.

See [token::Secret] for serialization examples

expiry_duration: Duration

Expiry duration of tokens, in seconds.

Defaults to 24 hours when deserialized and left unfilled

refresh_token: Option<RefreshTokenConfiguration>

Customise refresh token options. Set to None to disable refresh tokens

Methods

impl Configuration

pub fn refresh_token_enabled(&self) -> bool

Returns whether refresh tokens are enabled

pub fn refresh_token(&self) -> &RefreshTokenConfiguration

Convenience function to return a reference to the Refresh Token configuration.

Panics

Panics if refresh token is not enabled

pub fn keys(&self) -> Result<Keys, Error>

Prepare the keys for use with various cryptographic operations