Struct rowdy::token::Keys

pub struct Keys {
    pub signing: Secret,
    pub signature_verification: Secret,
    pub encryption: Option<JWK<Empty>>,
    pub decryption: Option<JWK<Empty>>,
}

Keys prepared in a form directly usable for cryptographic operations. This prevents us from having to repeatedly read keys from the file system. Users should prepare the keys from Configuration using Configuration::keys() and then use this struct to retrieve keys from instead of the functions from Secret.

Fields
signing: Secret

Key used to signed tokens

signature_verification: Secret

Key used to verify token signatures

encryption: Option<JWK<Empty>>

Key used to encrypt tokens. Used if Refresh tokens are enabled.

decryption: Option<JWK<Empty>>

Key used to decrypt tokens. Used if Refresh tokens are enabled.