Struct rowdy::token::Token

pub struct Token<T> {
    pub token: JWT<T, Empty>,
    pub expires_in: Duration,
    pub issued_at: DateTime<Utc>,
    pub refresh_token: Option<RefreshToken>,
}

A token that will be serialized into JSON and passed to clients. This encapsulates a JSON Web Token or JWT. Clients will pass the encapsulated JWT to services that require it. The JWT should be considered opaque to clients. The Token struct contains enough information for the client to act on, including expiry times.

Fields
token: JWT<T, Empty>

Tne encapsulated JWT.

expires_in: Duration

The duration from issued_at where the token will expire

issued_at: DateTime<Utc>

Time the token was issued at

refresh_token: Option<RefreshToken>

Refresh token, if enabled and requested for

Methods

impl<T: Serialize + DeserializeOwned + 'static> Token<T>

pub fn with_configuration(
config: &Configuration,
subject: &str,
service: &str,
private_claims: T,
refresh_token_payload: Option<&JsonValue>
) -> Result<Self, Error>

Based on the configuration, make a token for the subject, along with some private claims.

pub fn encode(self, secret: &Secret) -> Result<Self, Error>

Consumes self and encode the embedded JWT with signature. If the JWT is already encoded, this returns an error

pub fn decode(
self,
secret: &Secret,
algorithm: SignatureAlgorithm
) -> Result<Self, Error>

Consumes self and decode the embedded JWT with signature verification If the JWT is already decoded, this returns an error

pub fn is_decoded(&self) -> bool

Returns whether the wrapped token is decoded and verified

pub fn is_encoded(&self) -> bool

Returns whether the wrapped token is encoded and signed

pub fn registered_claims(&self) -> Result<&RegisteredClaims, Error>

Convenience function to extract the registered claims from a decoded token

pub fn private_claims(&self) -> Result<&T, Error>

Conveneince function to extract the private claims from a decoded token

pub fn header(&self) -> Result<&Header<Empty>, Error>

Convenience function to extract the headers from a decoded token

pub fn encoded_token(&self) -> Result<String, Error>

Convenience method to extract the encoded token

pub fn refresh_token(&self) -> Option<&RefreshToken>

Convenience method to obtain a reference to the refresh token

pub fn encrypt_refresh_token(
self,
secret: &Secret,
key: &JWK<Empty>
) -> Result<Self, Error>

Consumes self, and encrypt and sign the embedded refresh token

pub fn decrypt_refresh_token(
self,
secret: &Secret,
key: &JWK<Empty>,
signing_algorithm: SignatureAlgorithm,
cek_algorithm: KeyManagementAlgorithm,
enc_algorithm: ContentEncryptionAlgorithm
) -> Result<Self, Error>

Consumes self, and decrypt and verify the signature of the embedded refresh token

pub fn has_refresh_token(&self) -> bool

Returns whether there is a refresh token