Struct rowdy::auth::LdapAuthenticator

pub struct LdapAuthenticator {
    pub address: String,
    pub bind_dn: String,
    pub bind_password: String,
    pub search_base: String,
    pub search_filter: Option<String>,
    pub include_attributes: Vec<String>,
    pub attributes_namespace: Option<String>,
    pub subject_attribute: Option<String>,
}

LDAP based authenticator

Use LDAP server as the identity provider.

Example

use rowdy::auth::LdapAuthenticator;
let authenticator = LdapAuthenticator {
    address: "ldap://ldap.forumsys.com".to_string(),
    bind_dn: "cn=read-only-admin,dc=example,dc=com".to_string(),
    bind_password: "password".to_string(),
    search_base: "dc=example,dc=com".to_string(),
    search_filter: Some("(uid={account})".to_string()),
    include_attributes: vec!["cn".to_string()],
    attributes_namespace: Some("user".to_string()),
    subject_attribute: Some("uid".to_string()),
};

Fields
address: String

Location of the LDAP server

bind_dn: String

The user that we will bind to LDAP to search for users

bind_password: String

The password that we will use to bind to LDAP to search for users

search_base: String

Base to use when searching for user. {account} is expanded to the user's account. Search filters MUST be escaped according to RFC 4515.

search_filter: Option<String>

Filter to use when searching for user. {account} is expanded to the user's account. Search filters MUST be escaped according to RFC 4515.

include_attributes: Vec<String>

List of attributes from the LDAP Search Result Entry to be included in the JWT. The values will be placed under the attributes_namespace key in the JWT. Missing keys are silently ignored.

When a search with LDAP is performed, a SearchEntry is returned. You can use this to configure the list of attributes that should be included in the JSON Web Token that will be provided to your applications.

attributes_namespace: Option<String>

Namespace or Key in the returned JSON Web Token to return include_attributes attributes in. See the documentation for the include_attributes for more information.

If set to None, the keys from the LDAP attributes will be merged with that of the JWT, and any duplicate keys will result in errors as described here.

subject_attribute: Option<String>

The LDAP attribute to be used as the "subject" of the JWT token issued. By default, the dn attribute will be used. Another common attribute would be cn.

The first value returned by the attribute will be used as the subject.

Methods

impl LdapAuthenticator

pub fn connect(&self) -> Result<LdapConn, Error>

Connects to the LDAP server

pub fn bind(
&self,
connection: &LdapConn,
dn: &str,
password: &str
) -> Result<(), Error>

Bind a connection to some dn

Example

use rowdy::auth::LdapAuthenticator;
let authenticator = LdapAuthenticator {
    address: "ldap://ldap.forumsys.com".to_string(),
    bind_dn: "cn=read-only-admin,dc=example,dc=com".to_string(),
    bind_password: "password".to_string(),
    search_base: "dc=example,dc=com".to_string(),
    search_filter: Some("(uid={account})".to_string()),
    include_attributes: vec!["cn".to_string()],
    attributes_namespace: Some("user".to_string()),
    subject_attribute: Some("uid".to_string()),
};
let connection = authenticator.connect().unwrap();
authenticator.bind(&connection, "cn=read-only-admin,dc=example,dc=com", "password")
    .expect("To bind successfully");

pub fn verify(
&self,
username: &str,
password: &str,
include_refresh_payload: bool
) -> Result<AuthenticationResult, Error>

Authenticate the user with the username/password