Struct rowdy::token::RefreshTokenConfiguration

pub struct RefreshTokenConfiguration {
    pub cek_algorithm: KeyManagementAlgorithm,
    pub enc_algorithm: ContentEncryptionAlgorithm,
    pub key: Secret,
    pub expiry_duration: Duration,
}

Configuration for Refresh Tokens

Refresh Tokens are encrypted JWS, signed with the same algorithm as access tokens. There are two algorithms used.

A content encryption algorithm is used to encrypt the payload of the token, and provided some integrity protection. The algorithm used is symmetric. The list of supported algorithm can be found here. The key used to encrypt the content is called the Content Encryption Key (CEK).

Another algorithm is employed to determine and/or encrypt the CEK. The list of supported algorithms can be found here.

Fields
cek_algorithm: KeyManagementAlgorithm

Algorithm used to determine and/or encrypt the CEK

enc_algorithm: ContentEncryptionAlgorithm

Algorithm used to encrypt the content using the CEK

key: Secret

Key used in determining the CEK, or directly encrypt the content depending on the cek_algorithm

expiry_duration: Duration

Expiry duration of refresh tokens, in seconds.

Defaults to 24 hours when deserialized and left unfilled